requirements.txt
PyYAML==6.0
requests~=2.28.1
paramiko~=2.11.0